Refactor renderer globals code into new file
